fix digest test vectors test
this installs a workaround in `mc-crypto-digestible` for handling
of bytes primitive.

A better fix may be to implement `allow_omit` for the bytes primitive,
that would bring us more into line with protobuf schema evolution.

We're not sure yet if that would break our current hashes, needs
follow up work. If so then we can drop the `omit = "empty"` thing
added here.