Refactor rct bulletproofs struct, per Jcape suggestion
It was suggested that instead of having a bunch of repeated fields,
that all have to have the same length, we should just have one struct
that is repeated, and deprecate all the older repeated fields.

This is an experimental commit refactoring rct bulletproofs object
in this direction.

Right now, I think the change is a good idea in the long term, but in
the short term it is high risk because it makes this code harder to review.
It creates a bunch of additional churn around checking `RingMLSAG` signatures
due the need to support backwards compatibility

For example, in the `fn key_images(&self)` function, if we overlooked
that `RingMLSAG` can appear in two different places now, then we can
cause an infinite spending bug.

It might be simpler to just create an entirely new SignatureRctBulletproofs
object, than to keep nursing the old one along and branching everywhere
on block version. But this will definitely increase the complexity of review.
The only complexity being saved here is, avoiding the need to check that some
lists have the same length, which while annoying, isn't that complicated.