Improve parallel processing when there is only a single monitor
The goal of this is to be able to saturate all CPU cores with view
key scanning operations when there is a single monitor, which is
a pretty common case. In this case, we need to be able to scan
multiple blocks in parallel for a single monitor, when it is
many blocks behind.

I'm tired of waiting for mobilecoind to sync. :)